+++ taplo-ide/vscode/src/syntax/composite/array.ts
@@ -1,0 +1,23 @@
+export const array = {
+  begin: "(?<!\\w)(\\[)\\s*",
+  name: "meta.array.toml",
+  beginCaptures: {
+    1: {
+      name: "punctuation.definition.array.toml",
+    },
+  },
+  end: "\\s*(\\])(?!\\w)",
+  endCaptures: {
+    1: {
+      name: "punctuation.definition.array.toml",
+    },
+  },
+  patterns: [
+    {
+      include: "#comment",
+    },
+    {
+      include: "#value",
+    },
+  ],
+};