+++ taplo-ide/vscode/src/syntax/literal/boolean.ts
@@ -1,0 +1,12 @@
+export const boolean = [
+  {
+    match: "(?<!\\w)(true|false)(?!\\w)",
+    captures: {
+      1: {
+        name: "constant.language.boolean.toml",
+      },
+    },
+  },
+];
+
+