+++ taplo-ide/vscode/src/syntax/literal/index.ts
@@ -1,0 +1,5 @@
+export * from "./boolean";
+export * from "./number";
+export * from "./string";
+export * from "./datetime";
+